requirements.txt
streamlit
openai
pillow
pandas